# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing Wandersteine (hiking stones)",
        "title": "Trailmarks API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.trailmarks.io/support",
            "email": "support@trailmarks.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/wandersteine": {
            "get": {
                "description": "Get all hiking stones",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wandersteine"
                ],
                "summary": "Get all Wandersteine",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.WandersteinResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/wandersteine/recent": {
            "get": {
                "description": "Get the 5 most recently added hiking stones",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wandersteine"
                ],
                "summary": "Get recent Wandersteine",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.WandersteinResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.WandersteinResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Steiniger Pfad"
                },
                "preview_url": {
                    "type": "string",
                    "example": "https://example.com/images/wanderstein1.jpg"
                },
                "unique_id": {
                    "type": "string",
                    "example": "WS-2024-001"
                }
            }
        }
    }
}